services/bf5e5471be0f1ff7cd3844b8/requirements.txt
colorama==0.4.6
openai==0.27.2
python-dotenv==1.0.0
tiktoken==0.4.0
prettytable==3.7.0
PyMySQL==1.0.3
langchain==0.0.162
SQLAlchemy==1.4.39
numpy==1.24.3
pandas==1.5.3
cryptography==41.0.0
fastapi==0.95.0
uvicorn==0.22.0
jinja2==3.1.2

